Organize code into separate .cpp and .h files
