+++ OrderType.h
@@ -1,0 +1,10 @@
+#pragma once
+
+// GTC, F&K
+enum class OrderType
+{
+	GoodTillCancel,
+	FillAndKill,
+	FillOrKill,
+	GoodForDay,
+};
