+++ Side.h
@@ -1,0 +1,8 @@
+#pragma once
+
+// Buy, Sell
+enum class Side
+{
+	Buy,
+	Sell
+};
