+++ TradeInfo.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "Aliases.h"
+
+struct TradeInfo
+{
+	OrderId orderId_;
+	Price price_;
+	Quantity quantity_;
+};
